requirements.txt
bert-tensorflow
pandas
jieba
tensorflow_hub
